requirements.txt
pytest==9.0.2
allure-pytest==2.15.3
requests==2.32.5
pydantic[email]
python-dotenv~=1.2.1